Extracted module to be Android agnostic so can be used in any Java project.
