+++ library/src/main/java/com/squareup/otto/Produce.java
@@ -22,10 +22,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marks a method as an instance producer, as used by {@link AnnotatedHandlerFinder} and {@link Bus}.
+ * Marks a method as an instance producer, as used by {@link com.squareup.otto.vanilla.AnnotatedHandlerFinder}
+ * and {@link com.squareup.otto.vanilla.Bus}.
  * <p>
  * Otto infers the instance type from the annotated method's return type. Producer methods may return null when there is
- * no appropriate value to share. The calling {@link Bus} ignores such returns and posts nothing.
+ * no appropriate value to share. The calling {@link com.squareup.otto.vanilla.Bus} ignores such returns
+ * and posts nothing.
  *
  * @author Jake Wharton
  */
