+++ library/src/main/java/com/squareup/otto/Subscribe.java
@@ -22,12 +22,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marks a method as an event handler, as used by {@link AnnotatedHandlerFinder} and {@link Bus}.
+ * Marks a method as an event handler, as used by {@link com.squareup.otto.vanilla.AnnotatedHandlerFinder}
+ * and {@link com.squareup.otto.vanilla.Bus}.
  *
  * <p>The method's first (and only) parameter defines the event type.
  * <p>If this annotation is applied to methods with zero parameters or more than one parameter, the object containing
- * the method will not be able to register for event delivery from the {@link Bus}. Otto fails fast by throwing
- * runtime exceptions in these cases.
+ * the method will not be able to register for event delivery from the {@link com.squareup.otto.vanilla.Bus}.
+ * Otto fails fast by throwing runtime exceptions in these cases.
  *
  * @author Cliff Biffle
  */
