+++ library/src/main/java/com/squareup/otto/vanilla/EventHandler.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.squareup.otto;
+package com.squareup.otto.vanilla;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
