+++ library/src/test/java/com/squareup/otto/vanilla/LazyStringProducer.java
@@ -1,4 +1,6 @@
-package com.squareup.otto;
+package com.squareup.otto.vanilla;
+
+import com.squareup.otto.Produce;
 
 public class LazyStringProducer {
   public String value = null;
