+++ library/src/test/java/com/squareup/otto/vanilla/StringProducer.java
@@ -1,4 +1,6 @@
-package com.squareup.otto;
+package com.squareup.otto.vanilla;
+
+import com.squareup.otto.Produce;
 
 public class StringProducer {
   public static final String VALUE = "Hello, Producer";
