+++ sample/src/main/java/com/squareup/otto/sample/BusProvider.java
@@ -16,7 +16,7 @@
 
 package com.squareup.otto.sample;
 
-import com.squareup.otto.Bus;
+import com.squareup.otto.vanilla.Bus;
 
 /**
  * Maintains a singleton instance for obtaining the bus. Ideally this would be replaced with a more efficient means
